[#6] refactor : 폴더 구조 리팩토링
